improvements to dist and related
* improve dist to:
1. better work with tags
2. move dist db to separated file allows connect to dist db by others
3. don't show machines on submission if master node is storage only it confuses, but use global tags from all nodes
4. more cool stuff :)
